bpf: Factor out emission of ModR/M byte for *(dst_reg + off)
To be clear, there remains other code in the file that emits a ModR/M
byte. This patch only refactors a specific case.

I originally added this patch because I expected to add more uses of
the duplicated code. That turned out to be unnecessary but I think
this patch still makes the code I'm modifying easier to read.

Signed-off-by: Brendan Jackman <jackmanb@google.com>