bpf: Add instructions for atomic_cmpxchg and friends
These are the operations that implement atomic exchange and
compare-exchange.

They are peculiarly named because of the presence of the separate
FETCH field that tells you whether the instruction writes the value
back to the src register.

BPF_CMP is supported without BPF_FETCH, providing atomic_set. However
BPF_CMPSET without BPF_FETCH (i.e. an atomic compare-and-set without
knowing whether the write was successfuly) isn't implemented by the
kernel, x86, or ARM. It would be a burden on the JIT and it's hard to
imagine a use for this operation, so it's not supported.

There are two significant design decisions made for the CMPSET
instruction:

 - To solve the issue that this operation fundamentally has 3
   operands, but we only have two register fields. Therefore the
   operand we compare against (the kernel's API calls it 'old') is
   hard-coded to be R0. x86 has similar design (and A64 doesn't
   have this problem).

   A potential alternative might be to encode the other operand's
   register number in the immediate field.

 - The kernel's atomic_cmpxchg returns the old value, while the C11
   userspace APIs return a boolean indicating the comparison
   result. Which should BPF do? A64 returns the old value. x86 returns
   the old value in the hard-coded register (and also sets a
   flag). That means return-old-value is easier to JIT.

Signed-off-by: Brendan Jackman <jackmanb@google.com>
Change-Id: I3f19ad867dfd08515eecf72674e6fdefe28424bb