mm,memory_hotplug: Remove un-taken lock
It seems that [1] was acked, and the a v2 was written[2] which
improved upon it, but got bogged down in discussion of other topics,
so the improvements were not included. Then [1] got merged as commit
27cacaad16c ("mm,memory_hotplug: drop unneeded locking") and we ended
up with locks that get taken for read but never for write.

It's not really clear to me that the code I'm proposing here is
correct. However, it is not more wrong than before, and it is
definitely smaller!

Note that compared to Oscar's original v2[2], I have added a READ_ONCE
in page_outside_zone_boundaries; this is a substitute for the compiler
barrier that was implied by read_seqretry(). I don't think there's any
point in replacing the CPU read barrier because there is no
corresponding write barrier.

TODO: move this to a different commit?
TODO: run some selftests with KCSAN?

Theoretically scarier: it also seems possible under load/store tearing
to get values that have never been correct. E.g. I think if you race
with offline_pages, and the upper bytes of the spanned_pages that you
observe are older than the lower bytes, you could see a zone_end_pfn
that is higher than any PFN that was ever in the zone, maybe even one
that's spanned by another zone.

[1] https://lore.kernel.org/all/20210531093958.15021-1-osalvador@suse.de/T/#u
[2] https://lore.kernel.org/linux-mm/20210602091457.17772-3-osalvador@suse.de/#t

Co-developed-by: Oscar Salvador <osalvador@suse.de>
Signed-off-by: Brendan Jackman <jackmanb@google.com>