allow tests to pass in their on zonelist
I wanted to be able to inject NODE_DATA, but it turns out that this is
kind of a nop on !CONFIG_NUMA. CONFIG_NUMA is only defined for x86
(weird, I thought there were Arm NUMA systems) and I don't want to put
this in the x86 tree.

Maybe instead we could still go that route of just completely supporting
overriding NODE_DATA , not sure yet but for now I want to explore the
approach that Michal described where we create a fake node for the
tests.

EDIT: Yes actually we really do need to fake out NODE_DATA, because
that's used by page_node, which is used everywhere.

Next step is to figure out how to isolation (or even hotplug out) some
memory and put it into a fake node.