kunit: tool: Build compile_commands.json
compile_commands.json is used by clangd[1] to provide code navigation and
completion functionality to editors. See [2] for an example configuration that
includes this functionality for VSCode.

It can currently be built manually when using kunit.py, by running:

  ./scripts/clang-tools/gen_compile_commands.py -d .kunit

With this change however, it's built automatically so you don't need to
manually keep it up to date.

Unlike the manual approach, having make build the compile_commands.json means
that it appears in the build output tree instead of at the root of the source
tree, so you'll need to add --compile-commands-dir= to your clangd args for it
to be found.

[1] https://clangd.llvm.org/
[2] https://github.com/FlorentRevest/linux-kernel-vscode

Signed-off-by: Brendan Jackman <jackmanb@google.com>