mm,memory_hotplug: {READ,WRITE}_ONCE unsynchronized zone data
These fields are written by memory hotplug under mem_hotplug_lock but
read without any lock. I assume that all of this reader code is robust
against the value being stale or "from the future", but we also need
to account for:

1. Load/store tearing (TBH I suspect this one is purely theoretical).

2. Undefined behaviour - unless I'm mistaken the compiler is free to
   spill and re-read these fields and assume that they have not
   changed. This one feels like a more practical concern.

As I wrote above, I "assume" that this code is robust against
unordered changes to these values... I don't know if the code I'm
proposing here is correct but if not I believe it's closer to correct
than before.

Signed-off-by: Brendan Jackman <jackmanb@google.com>